Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float N;
-    scanf("%f\n",&N);
-    printf("You entered: %.4f",N);
+    double N;
+    scanf("%lf\n",&N);
+    printf("You entered: %.4lf",N);
     return 0;
 }
